+++ Documento.cs
@@ -9,7 +9,7 @@ namespace csharp_biblioteca_db
     enum Stato { Disponibile, Prestito }
     internal class Documento
     {
-        public int Codice { get; set; }
+        public long Codice { get; set; }
         public string Titolo { get; set; }
         public int Anno { get; set; }
         public string Settore { get; set; }
@@ -17,7 +17,7 @@ namespace csharp_biblioteca_db
         public List<Autore> Autori { get; set; }
         public Scaffale Scaffale { get; set; }
 
-        public Documento(int Codice, string Titolo, int Anno, string Settore, string sScaffale)
+        public Documento(long Codice, string Titolo, int Anno, string Settore, string sScaffale)
         {
             this.Codice = Codice;
             this.Titolo = Titolo;
